requirements.txt
decorator==4.3.2
Django==2.1.7
djangorestframework==3.9.1
Faker==1.0.2
Markdown==3.0.1
mysqlclient==1.4.2.post1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
text-unidecode==1.2
